+++ python/adjoint/filters.py
@@ -3,10 +3,13 @@ A collection of routines for use in topology optimization comprising
 convolution filters (kernels), projection operators, and morphological
 transforms.
 """
+
+import sys
+from typing import List, Tuple, Union
+
 import numpy as np
 from autograd import numpy as npa
 from scipy import signal, special
-from typing import List, Tuple, Union
 
 ArrayLikeType = Union[List, Tuple, np.ndarray]
 
@@ -248,13 +251,13 @@ def mesh_grid(
         periodic_axes = np.array(periodic_axes)
         if 0 in periodic_axes:
             xv = (
-                np.arange(0, np.ceil(2 * radius / Lx) * Lx / 2, 1 / resolution[0])
+                npa.arange(0, npa.ceil(2 * radius / Lx) * Lx / 2, 1 / resolution[0])
                 if resolution[0] > 0
                 else [0]
             )
         if 1 in periodic_axes:
             yv = (
-                np.arange(0, np.ceil(2 * radius / Ly) * Ly / 2, 1 / resolution[1])
+                npa.arange(0, npa.ceil(2 * radius / Ly) * Ly / 2, 1 / resolution[1])
                 if resolution[1] > 0
                 else [0]
             )
@@ -327,7 +330,7 @@ def conic_filter(
     """
     Nx, Ny, X, Y = mesh_grid(radius, Lx, Ly, resolution, periodic_axes)
     x = x.reshape(Nx, Ny)  # Ensure the input is 2d
-    h = np.where(
+    h = npa.where(
         X**2 + Y**2 < radius**2, (1 - np.sqrt(abs(X**2 + Y**2)) / radius), 0
     )
     return convolve_design_weights_and_kernel(x, h, periodic_axes)
@@ -694,9 +697,143 @@ def tanh_projection(x: np.ndarray, beta: float, eta: float) -> np.ndarray:
     Returns:
         The filtered design weights.
     """
+    if beta == npa.inf:
+        # Note that backpropagating through here can produce NaNs. So we
+        # manually specify the step function to keep the gradient clean.
+        return npa.where(x > eta, 1.0, 0.0)
+    else:
+        return (npa.tanh(beta * eta) + npa.tanh(beta * (x - eta))) / (
+            npa.tanh(beta * eta) + npa.tanh(beta * (1 - eta))
+        )
 
-    return (npa.tanh(beta * eta) + npa.tanh(beta * (x - eta))) / (
-        npa.tanh(beta * eta) + npa.tanh(beta * (1 - eta))
+
+def smoothed_projection(
+    rho_filtered: ArrayLikeType,
+    beta: float,
+    eta: float,
+    resolution: float,
+):
+    """Project using subpixel smoothing, which allows for β→∞.
+
+    This technique integrates out the discontinuity within the projection
+    function, allowing the user to smoothly increase β from 0 to ∞ without
+    losing the gradient. Effectively, a level set is created, and from this
+    level set, first-order subpixel smoothing is applied to the interfaces (if
+    any are present).
+
+    In order for this to work, the input array must already be smooth (e.g. by
+    filtering).
+
+    While the original approach involves numerical quadrature, this approach
+    performs a "trick" by assuming that the user is always infinitely projecting
+    (β=∞). In this case, the expensive quadrature simplifies to an analytic
+    fill-factor expression. When to use this fill factor requires some careful
+    logic.
+
+    For one, we want to make sure that the user can indeed project at any level
+    (not just infinity). So in these cases, we simply check if in interface is
+    within the pixel. If not, we revert to the standard filter plus project
+    technique.
+
+    If there is an interface, we want to make sure the derivative remains
+    continuous both as the interface leaves the cell, *and* as it crosses the
+    center. To ensure this, we need to account for the different possibilities.
+
+    Args:
+        rho_filtered: The (2D) input design parameters (already filered e.g.
+            with a conic filter).
+        beta: The thresholding parameter in the range [0, inf]. Determines the
+            degree of binarization of the output.
+        eta: The threshold point in the range [0, 1].
+        resolution: resolution of the design grid (not the Meep grid
+            resolution).
+    Returns:
+        The projected and smoothed output.
+
+    Example:
+        >>> Lx = 2; Ly = 2
+        >>> resolution = 50
+        >>> eta_i = 0.5; eta_e = 0.75
+        >>> lengthscale = 0.1
+        >>> filter_radius = get_conic_radius_from_eta_e(lengthscale, eta_e)
+        >>> Nx = onp.round(Lx * resolution) + 1
+        >>> Ny = onp.round(Ly * resolution) + 1
+        >>> rho = onp.random.rand(Nx, Ny)
+        >>> beta = npa.inf
+        >>> rho_filtered = conic_filter(rho, filter_radius, Lx, Ly, resolution)
+        >>> rho_projected = smoothed_projection(rho_filtered, beta, eta_i, resolution)
+    """
+    # Note that currently, the underlying assumption is that the smoothing
+    # kernel is a circle, which means dx = dy.
+    dx = dy = 1 / resolution
+    R_smoothing = 0.55 * dx
+
+    rho_projected = tanh_projection(rho_filtered, beta=beta, eta=eta)
+
+    # Compute the spatial gradient (using finite differences) of the *filtered*
+    # field, which will always be smooth and is the key to our approach. This
+    # gradient essentially represents the normal direction pointing the the
+    # nearest inteface.
+    rho_filtered_grad = npa.gradient(rho_filtered)
+    rho_filtered_grad_helper = (rho_filtered_grad[0] / dx) ** 2 + (
+        rho_filtered_grad[1] / dy
+    ) ** 2
+
+    # Note that a uniform field (norm=0) is problematic, because it creates
+    # divide by zero issues and makes backpropagation difficult, so we sanitize
+    # and determine where smoothing is actually needed. The value where we don't
+    # need smoothings doesn't actually matter, since all our computations our
+    # purely element-wise (no spatial locality) and those pixels will instead
+    # rely on the standard projection. So just use 1, since it's well behaved.
+    nonzero_norm = npa.abs(rho_filtered_grad_helper) > 0
+
+    rho_filtered_grad_norm = npa.sqrt(
+        npa.where(nonzero_norm, rho_filtered_grad_helper, 1)
+    )
+    rho_filtered_grad_norm_eff = npa.where(nonzero_norm, rho_filtered_grad_norm, 1)
+
+    # The distance for the center of the pixel to the nearest interface
+    d = (eta - rho_filtered) / rho_filtered_grad_norm_eff
+
+    # Only need smoothing if an interface lies within the voxel. Since d is
+    # actually an "effective" d by this point, we need to ignore values that may
+    # have been sanitized earlier on.
+    needs_smoothing = nonzero_norm & (npa.abs(d) < R_smoothing)
+
+    # The fill factor is used to perform simple, first-order subpixel smoothing.
+    # We use the (2D) analytic expression that comes when assuming the smoothing
+    # kernel is a circle. Note that because the kernel contains some
+    # expressions that are sensitive to NaNs, we have to use the "double where"
+    # trick to avoid the Nans in the backward trace. This is a common problem
+    # with array-based AD tracers, apparently. See here:
+    # https://github.com/google/jax/issues/1052#issuecomment-5140833520
+    d_R = d / R_smoothing
+    F = npa.where(
+        needs_smoothing, 0.5 - 15 / 16 * d_R + 5 / 8 * d_R**3 - 3 / 16 * d_R**5, 1.0
+    )
+    # F(-d)
+    F_minus = npa.where(
+        needs_smoothing, 0.5 + 15 / 16 * d_R - 5 / 8 * d_R**3 + 3 / 16 * d_R**5, 1.0
+    )
+
+    # Determine the upper and lower bounds of materials in the current pixel (before projection).
+    rho_filtered_minus = rho_filtered - R_smoothing * rho_filtered_grad_norm_eff * F
+    rho_filtered_plus = (
+        rho_filtered + R_smoothing * rho_filtered_grad_norm_eff * F_minus
+    )
+
+    # Finally, we project the extents of our range.
+    rho_minus_eff_projected = tanh_projection(rho_filtered_minus, beta=beta, eta=eta)
+    rho_plus_eff_projected = tanh_projection(rho_filtered_plus, beta=beta, eta=eta)
+
+    # Only apply smoothing to interfaces
+    rho_projected_smoothed = (
+        1 - F
+    ) * rho_minus_eff_projected + F * rho_plus_eff_projected
+    return npa.where(
+        needs_smoothing,
+        rho_projected_smoothed,
+        rho_projected,
     )