+++ src/fields.cpp
@@ -1,4 +1,4 @@
-/* Copyright (C) 2005-2023 Massachusetts Institute of Technology
+/* Copyright (C) 2005-2025 Massachusetts Institute of Technology
 %
 %  This program is free software; you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -30,10 +30,10 @@ using namespace std;
 namespace meep {
 
 fields::fields(structure *s, double m, double beta, bool zero_fields_near_cylorigin,
-               int loop_tile_base_db, int loop_tile_base_eh)
+               int loop_tile_base_db, int loop_tile_base_eh, std::vector<double> bfast_scaled_k)
     : S(s->S), gv(s->gv), user_volume(s->user_volume), v(s->v), m(m), beta(beta),
       loop_tile_base_db(loop_tile_base_db), loop_tile_base_eh(loop_tile_base_eh),
-      working_on(&times_spent) {
+      working_on(&times_spent), bfast_scaled_k(bfast_scaled_k) {
   shared_chunks = s->shared_chunks;
   components_allocated = false;
   synchronized_magnetic_fields = 0;
@@ -60,7 +60,7 @@ fields::fields(structure *s, double m, double beta, bool zero_fields_near_cylori
   chunks = new fields_chunk_ptr[num_chunks];
   for (int i = 0; i < num_chunks; i++)
     chunks[i] = new fields_chunk(s->chunks[i], outdir, m, beta, zero_fields_near_cylorigin, i,
-                                 loop_tile_base_db);
+                                 loop_tile_base_db, bfast_scaled_k);
   FOR_FIELD_TYPES(ft) {
     typedef realnum *realnum_ptr;
     comm_blocks[ft] = new realnum_ptr[num_chunks * num_chunks];
@@ -93,6 +93,7 @@ fields::fields(const fields &thef)
   outdir = new char[strlen(thef.outdir) + 1];
   strcpy(outdir, thef.outdir);
   m = thef.m;
+  bfast_scaled_k = thef.bfast_scaled_k;
   beta = thef.beta;
   phasein_time = thef.phasein_time;
   for (int d = 0; d < 5; d++) {
@@ -171,12 +172,14 @@ fields_chunk::~fields_chunk() {
     delete[] f_u[c][cmp];
     delete[] f_w[c][cmp];
     delete[] f_cond[c][cmp];
+    delete[] f_bfast[c][cmp];
     delete[] f_minus_p[c][cmp];
     delete[] f_w_prev[c][cmp];
     delete[] f_backup[c][cmp];
     delete[] f_u_backup[c][cmp];
     delete[] f_w_backup[c][cmp];
     delete[] f_cond_backup[c][cmp];
+    delete[] f_bfast_backup[c][cmp];
   }
   delete[] f_rderiv_int;
   while (dft_chunks) {
@@ -237,9 +240,10 @@ void check_tiles(grid_volume gv, const std::vector<grid_volume> &gvs) {
 }
 
 fields_chunk::fields_chunk(structure_chunk *the_s, const char *od, double m, double beta,
-                           bool zero_fields_near_cylorigin, int chunkidx, int loop_tile_base_db)
+                           bool zero_fields_near_cylorigin, int chunkidx, int loop_tile_base_db,
+                           std::vector<double> bfast_scaled_k)
     : gv(the_s->gv), v(the_s->v), m(m), zero_fields_near_cylorigin(zero_fields_near_cylorigin),
-      beta(beta) {
+      beta(beta), bfast_scaled_k(bfast_scaled_k) {
   s = the_s;
   chunk_idx = chunkidx;
   s->refcount++;
@@ -278,12 +282,14 @@ fields_chunk::fields_chunk(structure_chunk *the_s, const char *od, double m, dou
     f_u[c][cmp] = NULL;
     f_w[c][cmp] = NULL;
     f_cond[c][cmp] = NULL;
+    f_bfast[c][cmp] = NULL;
     f_minus_p[c][cmp] = NULL;
     f_w_prev[c][cmp] = NULL;
     f_backup[c][cmp] = NULL;
     f_u_backup[c][cmp] = NULL;
     f_w_backup[c][cmp] = NULL;
     f_cond_backup[c][cmp] = NULL;
+    f_bfast_backup[c][cmp] = NULL;
   }
   f_rderiv_int = NULL;
   FOR_FIELD_TYPES(ft) {
@@ -299,6 +305,7 @@ fields_chunk::fields_chunk(const fields_chunk &thef, int chunkidx) : gv(thef.gv)
   s->refcount++;
   outdir = thef.outdir;
   m = thef.m;
+  bfast_scaled_k = thef.bfast_scaled_k;
   zero_fields_near_cylorigin = thef.zero_fields_near_cylorigin;
   beta = thef.beta;
   new_s = thef.new_s;
@@ -333,10 +340,12 @@ fields_chunk::fields_chunk(const fields_chunk &thef, int chunkidx) : gv(thef.gv)
     f_u[c][cmp] = NULL;
     f_w[c][cmp] = NULL;
     f_cond[c][cmp] = NULL;
+    f_bfast[c][cmp] = NULL;
     f_backup[c][cmp] = NULL;
     f_u_backup[c][cmp] = NULL;
     f_w_backup[c][cmp] = NULL;
     f_cond_backup[c][cmp] = NULL;
+    f_bfast_backup[c][cmp] = NULL;
   }
   FOR_COMPONENTS(c) DOCMP {
     if (!is_magnetic(c) && thef.f[c][cmp]) {
@@ -354,6 +363,10 @@ fields_chunk::fields_chunk(const fields_chunk &thef, int chunkidx) : gv(thef.gv)
     if (thef.f_cond[c][cmp]) {
       f_cond[c][cmp] = new realnum[gv.ntot()];
       memcpy(f_cond[c][cmp], thef.f_cond[c][cmp], sizeof(realnum) * gv.ntot());
+    }
+    if (thef.f_bfast[c][cmp]) {
+      f_bfast[c][cmp] = new realnum[gv.ntot()];
+      memcpy(f_bfast[c][cmp], thef.f_bfast[c][cmp], sizeof(realnum) * gv.ntot());
     }
   }
   FOR_MAGNETIC_COMPONENTS(c) DOCMP {
@@ -617,10 +630,12 @@ void fields_chunk::zero_fields() {
     ZERO(f_u[c][cmp]);
     ZERO(f_w[c][cmp]);
     ZERO(f_cond[c][cmp]);
+    ZERO(f_bfast[c][cmp]);
     ZERO(f_backup[c][cmp]);
     ZERO(f_u_backup[c][cmp]);
     ZERO(f_w_backup[c][cmp]);
     ZERO(f_cond_backup[c][cmp]);
+    ZERO(f_bfast_backup[c][cmp]);
 #undef ZERO
   }
   if (is_mine()) FOR_FIELD_TYPES(ft) {